# docsclear/swagger.yaml
basePath: /
definitions:
...
  commissionratehandler.CommissionRateRequest:
    properties:
      is_active:
        example: true
        type: boolean
      max_commission_cents:
        example: 50000
        type: integer
      min_commission_cents:
        example: 10000
        type: integer
      rate_bps:
        example: 250
        type: integer
      shop_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      transaction_type:
        example: online_payment
        type: string
    type: object
...
  utils.AppError:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@goshop.dev
    name: API Support
  description: GoShop - E-commerce API with observability and metrics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: 1.0.0
paths:
...
  /api/admin/commission-rates:
    get:
      consumes:
      - application/json
      description: Retourne la liste de tous les taux de commission configurés pour
        une boutique donnée.
      parameters:
      - description: ID de la boutique (UUID)
        in: query
        name: shop_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: shop_id manquant
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Récupérer les taux de commission d'une boutique
      tags:
      - Commission Management
    put:
      consumes:
      - application/json
      description: Met à jour ou crée un nouveau taux de commission pour un type de
        transaction spécifique. Le taux est limité à 15% (1500 bps) conformément à
        la BCEAO.
      parameters:
      - description: Détails du taux de commission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/commissionratehandler.CommissionRateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Payload invalide ou taux hors limites
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Mettre à jour le taux de commission d'une boutique
      tags:
      - Commission Management
  /api/admin/commission-rates/trigger-credit:
    post:
      description: Lance la collecte des commissions pour les crédits en arrière-plan.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Déclencher manuellement la collecte des commissions Crédit
      tags:
      - Commission Management
  /api/admin/commission-rates/trigger-online:
    post:
      description: Lance la collecte des commissions pour les paiements en ligne en
        arrière-plan (utile pour le débogage ou la récupération).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Déclencher manuellement la collecte des commissions en ligne
      tags:
      - Commission Management
  /api/admin/commission-rates/trigger-tontine:
    post:
      description: Lance la collecte des commissions pour les tontines en arrière-plan.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erreur interne du serveur
          schema:
            $ref: '#/definitions/utils.AppError'
      security:
      - ApiKeyAuth: []
      summary: Déclencher manuellement la collecte des commissions Tontine
      tags:
      - Commission Management
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"