# interfaces/handler/commission_rate_handler/commission_rate_handler.go
package commissionratehandler

import (
	"context"
	"encoding/json"
	"net/http"

	appscheduler "Goshop/application/scheduler"
	"Goshop/domain/entity"
	"Goshop/domain/repository"
	"Goshop/interfaces/utils"

	"github.com/google/uuid"
	"github.com/rs/zerolog"
)

// ============================================================
// COMMISSION RATE HANDLER
// ============================================================

// CommissionRateHandler gère les endpoints de configuration des taux
// et les triggers manuels pour les schedulers
type CommissionRateHandler struct {
	rateRepo           repository.CommissionRateRepository
	onlinePaymentSched *appscheduler.OnlinePaymentScheduler
	tontineSched       *appscheduler.TontineScheduler
	creditSched        *appscheduler.CreditScheduler
}

// NewCommissionRateHandler crée une nouvelle instance
func NewCommissionRateHandler(
	rateRepo repository.CommissionRateRepository,
	onlinePaymentSched *appscheduler.OnlinePaymentScheduler,
	tontineSched *appscheduler.TontineScheduler,
	creditSched *appscheduler.CreditScheduler,
) *CommissionRateHandler {
	return &CommissionRateHandler{
		rateRepo:           rateRepo,
		onlinePaymentSched: onlinePaymentSched,
		tontineSched:       tontineSched,
		creditSched:        creditSched,
	}
}

// ============================================================
// REQUEST/RESPONSE TYPES
// ============================================================

// CommissionRateRequest représente la requête de mise à jour d'un taux de commission
type CommissionRateRequest struct {
	ShopID             string `json:"shop_id" example:"123e4567-e89b-12d3-a456-426614174000"`
	TransactionType    string `json:"transaction_type" example:"online_payment"`
	RateBps            int    `json:"rate_bps" example:"250"`
	MinCommissionCents int64  `json:"min_commission_cents" example:"10000"`
	MaxCommissionCents int64  `json:"max_commission_cents" example:"50000"`
	IsActive           bool   `json:"is_active" example:"true"`
}

// ============================================================
// ENDPOINTS : CONFIGURATION DES TAUX
// ============================================================

// @Summary Mettre à jour le taux de commission d'une boutique
// @Description Met à jour ou crée un nouveau taux de commission pour un type de transaction spécifique. Le taux est limité à 15% (1500 bps) conformément à la BCEAO.
// @Tags Commission Management
// @Accept json
// @Produce json
// @Param request body commissionratehandler.CommissionRateRequest true "Détails du taux de commission"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} utils.AppError "Payload invalide ou taux hors limites"
// @Failure 500 {object} utils.AppError "Erreur interne du serveur"
// @Security ApiKeyAuth
// @Router /api/admin/commission-rates [put]
func (h *CommissionRateHandler) UpdateRate(w http.ResponseWriter, r *http.Request) error {
	ctx := r.Context()
	logger := zerolog.Ctx(ctx)

	var req CommissionRateRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		return utils.NewAppError("INVALID_PAYLOAD", "Invalid request body", http.StatusBadRequest)
	}

	// Valider le type de transaction
	validTypes := map[string]bool{
		entity.TransactionTypeOnlinePayment: true,
		entity.TransactionTypeCOD:           true,
		"tontine_commercial":                true,
		"tontine_corporate":                 true,
		"tontine_family":                    true,
		entity.TransactionTypeCredit:        true,
	}

	if !validTypes[req.TransactionType] {
		return utils.NewAppError("INVALID_TRANSACTION_TYPE", "Invalid transaction type", http.StatusBadRequest)
	}

	// Valider le taux (max 15% = 1500 bps, BCEAO)
	if req.RateBps < 0 || req.RateBps > 1500 {
		return utils.NewAppError("INVALID_RATE", "Rate must be between 0 and 1500 bps (15%)", http.StatusBadRequest)
	}

	// Upsert
	rate := &entity.CommissionRate{
		ID:                 uuid.New().String(),
		ShopID:             req.ShopID,
		TransactionType:    req.TransactionType,
		RateBps:            req.RateBps,
		MinCommissionCents: req.MinCommissionCents,
		MaxCommissionCents: req.MaxCommissionCents,
		IsActive:           req.IsActive,
	}

	err := h.rateRepo.Update(ctx, rate)
	if err != nil {
		if err := h.rateRepo.Create(ctx, rate); err != nil {
			logger.Error().Err(err).Msg("Failed to upsert rate")
			return utils.NewAppError("RATE_UPSERT_FAILED", "Failed to update rate", http.StatusInternalServerError)
		}
	}

	logger.Info().
		Str("shop_id", req.ShopID).
		Str("transaction_type", req.TransactionType).
		Int("rate_bps", req.RateBps).
		Msg("Commission rate updated")

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"message": "Commission rate updated",
		"rate":    rate,
	})
	return nil
}

// @Summary Récupérer les taux de commission d'une boutique
// @Description Retourne la liste de tous les taux de commission configurés pour une boutique donnée.
// @Tags Commission Management
// @Accept json
// @Produce json
// @Param shop_id query string true "ID de la boutique (UUID)"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} utils.AppError "shop_id manquant"
// @Failure 500 {object} utils.AppError "Erreur interne du serveur"
// @Security ApiKeyAuth
// @Router /api/admin/commission-rates [get]
func (h *CommissionRateHandler) GetRates(w http.ResponseWriter, r *http.Request) error {
	ctx := r.Context()
	logger := zerolog.Ctx(ctx)

	shopID := r.URL.Query().Get("shop_id")
	if shopID == "" {
		return utils.NewAppError("MISSING_SHOP_ID", "shop_id is required", http.StatusBadRequest)
	}

	rates, err := h.rateRepo.FindByShop(ctx, shopID)
	if err != nil {
		logger.Error().Err(err).Msg("Failed to fetch rates")
		return utils.NewAppError("RATES_FETCH_FAILED", "Failed to fetch rates", http.StatusInternalServerError)
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"count":   len(rates),
		"rates":   rates,
	})
	return nil
}

// ============================================================
// ENDPOINTS : TRIGGERS MANUELS
// ============================================================

// @Summary Déclencher manuellement la collecte des commissions en ligne
// @Description Lance la collecte des commissions pour les paiements en ligne en arrière-plan (utile pour le débogage ou la récupération).
// @Tags Commission Management
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 500 {object} utils.AppError "Erreur interne du serveur"
// @Security ApiKeyAuth
// @Router /api/admin/commission-rates/trigger-online [post]
func (h *CommissionRateHandler) TriggerOnlineCollection(w http.ResponseWriter, r *http.Request) error {
	logger := zerolog.Ctx(r.Context())
	logger.Info().Msg("🔧 Manual online payment collection triggered")

	go func() {
		bgCtx := context.Background()
		if h.onlinePaymentSched != nil {
			if err := h.onlinePaymentSched.RunCollection(bgCtx); err != nil {
				logger.Error().Err(err).Msg("❌ Manual online payment collection failed")
			}
		}
	}()

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"message": "Online payment collection triggered in background",
	})
	return nil
}

// @Summary Déclencher manuellement la collecte des commissions Tontine
// @Description Lance la collecte des commissions pour les tontines en arrière-plan.
// @Tags Commission Management
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 500 {object} utils.AppError "Erreur interne du serveur"
// @Security ApiKeyAuth
// @Router /api/admin/commission-rates/trigger-tontine [post]
func (h *CommissionRateHandler) TriggerTontineCollection(w http.ResponseWriter, r *http.Request) error {
	logger := zerolog.Ctx(r.Context())
	logger.Info().Msg("🎯 Manual tontine collection triggered")

	go func() {
		bgCtx := context.Background()
		if h.tontineSched != nil {
			if err := h.tontineSched.RunCollection(bgCtx); err != nil {
				logger.Error().Err(err).Msg("❌ Manual tontine collection failed")
			}
		} else {
			logger.Warn().Msg("⚠️ Tontine scheduler not initialized")
		}
	}()

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"message": "Tontine collection triggered in background",
	})
	return nil
}

// @Summary Déclencher manuellement la collecte des commissions Crédit
// @Description Lance la collecte des commissions pour les crédits en arrière-plan.
// @Tags Commission Management
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 500 {object} utils.AppError "Erreur interne du serveur"
// @Security ApiKeyAuth
// @Router /api/admin/commission-rates/trigger-credit [post]
func (h *CommissionRateHandler) TriggerCreditCollection(w http.ResponseWriter, r *http.Request) error {
	logger := zerolog.Ctx(r.Context())
	logger.Info().Msg("💰 Manual credit collection triggered")

	go func() {
		bgCtx := context.Background()
		if h.creditSched != nil {
			if err := h.creditSched.RunCollection(bgCtx); err != nil {
				logger.Error().Err(err).Msg("❌ Manual credit collection failed")
			}
		} else {
			logger.Warn().Msg("⚠️ Credit scheduler not initialized")
		}
	}()

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{
		"success": true,
		"message": "Credit collection triggered in background",
	})
	return nil
}